+++ src/main/java/blackjack/domain/Gambler.java
@@ -1,0 +1,20 @@
+package blackjack.domain;
+
+import blackjack.domain.card.Card;
+import java.util.List;
+
+public abstract class Gambler {
+    protected final Hands hands = new Hands();
+
+    public void addCard(final Card card) {
+        hands.addNewCard(card);
+    }
+
+    public int calculateSum() {
+        return hands.calculateSum();
+    }
+
+    public List<Card> getCards() {
+        return hands.getCards();
+    }
+}
