+++ src/main/java/blackjack/domain/Player.java
@@ -1,12 +1,9 @@
 package blackjack.domain;
 
-import blackjack.domain.card.Card;
-import java.util.List;
 import java.util.Objects;
 
-public class Player {
+public class Player extends Gambler {
     private final String name;
-    private final Hands hands = new Hands();
 
     public Player(final String name) {
         validateName(name);
@@ -27,19 +24,7 @@ public class Player {
         return name;
     }
 
-    public void addCard(final Card card) {
-        hands.addNewCard(card);
-    }
-
-    public int calculateSum() {
-        return hands.calculateSum();
-    }
-
     public boolean isNameEquals(String playerName) {
         return Objects.equals(playerName, name);
-    }
-
-    public List<Card> getCards() {
-        return hands.getCards();
     }
 }
